Final rev
TODO (for next user study): Adjust logging to enable more streamlined
data analysis. It was time consuming to convert the numerous .db files
to .csv files for analysis. Adjust game difficulties, increasing the
range of difficulty levels (people did not notice a huge difference in
difficulty). Improve instructions for the users. Improve algorithm for
choosing the active side of the user (maybe incorporate a running
average velocity check as well as position?).  Any way to allow the use
of a limp limb?